Renamed AWindow, AChildWindow, APopupWindow, and AControlWindow to include suffix 'W', as well as added preprocessor macro switches for non-suffix use. This is to prepare for having Unicode and ANSI versions to stay in line with Windows design concepts.
AWindowW: Added conditional in create method to check if m_hWnd is NULL before calling CreateWindowEx. If m_hWnd is NOT NULL, we just return m_hWnd. This is to prevent overwriting the handle and having a window lost to the void.

AChildWindowW: Added conditional in create method to check if m_hWndParent is NULL before calling CreateWindowEx.

Added EditControlW class with non-suffix macro switch for future ANSI version.